refactor notification (not working)
