+++ JokeAbout/app/src/main/java/com/soldatov/jokeabout/presentation/JokeAboutActivity.kt
@@ -7,15 +7,18 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.RadioGroup
 import com.soldatov.jokeabout.R
+import com.soldatov.jokeabout.data.api.JokeApiImpl
 import com.soldatov.jokeabout.domain.models.Data
+import com.soldatov.jokeabout.presentation.utils.JokeNotification
 import com.soldatov.jokeabout.presentation.utils.UseCaseProvider
 
-class JokeAboutActivity : AppCompatActivity() {
+class JokeAboutActivity : AppCompatActivity(), JokeApiImpl.OnGetJoke {
 
     private lateinit var radioGroup: RadioGroup
     private lateinit var firstName: EditText
     private lateinit var lastName: EditText
     private lateinit var save: Button
+    private val jokeNotification = JokeNotification()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -46,5 +49,9 @@ class JokeAboutActivity : AppCompatActivity() {
             1 -> 60000
             else -> 3600000
         }
+    }
+
+    override fun onGetJoke(contentText: String) {
+        jokeNotification.showNotification(this, contentText)
     }
 }
